requirements.txt
scikit-learn>=1.4.0
scipy>=1.11.0
rapidfuzz>=3.9.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0.post1
numpy>=1.26.0
tqdm>=4.66.0
requests>=2.32.0
python-dotenv>=1.0.1
discord.py>=2.3.2
joblib>=1.3.0
# Tesztelés
pytest>=7.0.0           # Teszt keretrendszer :cite[3]
pytest-cov>=4.0.0       # Teszt lefedettség

# Code quality
flake8>=6.0.0           # Code style ellenőrzés :cite[3]
black>=23.0.0           # Kód formázás :cite[3]
mypy>=1.0.0             # Statikus típusellenőrzés :cite[3]

# Dokumentáció
sphinx>=7.0.0           # Dokumentáció generálás

# Biztonság
safety>=2.0.0           # Biztonsági rések ellenőrzése :cite[6]